+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.1] - 2019-07-11
 ### Added
 - Crossbows. These have some minor animation issues (because vanilla is dumb), but they function correctly.
 ### Changed
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Some issues with the JEI plugin on servers
 - Blended head colors including non-main parts
+- Possibly improves part analyzer performance
 
 ## [1.2.0] - 2019-07-01
 Updated for Minecraft 1.14.3